+++ apps/vite-demo/src/main.ts
@@ -1,7 +1,10 @@
 import { createApp } from 'vue'
 import './style.css'
 import App from './App.vue'
-import ButtonInstall from '@yolo/pg-components'
+import ButtonInstall from '@yolo/pg-components/button'
+import TagInstall from '@yolo/pg-components/tag'
 
-
-createApp(App).use(ButtonInstall).mount('#app')
+const app = createApp(App) 
+app.use(ButtonInstall)
+app.use(TagInstall)
+app.mount('#app')
