+++ packages/ui/pg-components/components/index.js
@@ -1,10 +1,2 @@
-import PgButton from './button/src/index.vue'
-
-const ButtonInstall = () => {
-  install: (app) => {
-    app.component('pg-button', PgButton);
-  }
-}
-
-export default ButtonInstall
-export { PgButton }
+export * from './button'
+export * from './tag'
