+++ packages/ui/pg-components/components/tag/index.js
@@ -1,0 +1,11 @@
+import PgTag from './src/index.vue'
+
+const TagInstall = {
+  install(app) {
+    app.component('pg-tag', PgTag)
+  }
+}
+
+export default TagInstall
+
+export { PgTag }
