+++ packages/ui/pg-components/components/tag/src/index.vue
@@ -1,0 +1,3 @@
+<template>
+  <div>我是tag</div>
+</template>
